+++ src/components/Alert/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-11-30: Initial version.
+- 2025-12-12: Update the styles - padding and gap - to make it look better.